chore (#65) :: DeleteAnnouncementService에 삭제 권한 검증 예외 처리 로직 추가

+++ src/main/java/team/jeonghokim/daedongyeojido/domain/announcement/service/DeleteAnnouncementService.java
@@ -5,7 +5,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import team.jeonghokim.daedongyeojido.domain.announcement.domain.Announcement;
 import team.jeonghokim.daedongyeojido.domain.announcement.domain.repository.AnnouncementRepository;
+import team.jeonghokim.daedongyeojido.domain.announcement.exception.AnnouncementAccessDeniedException;
 import team.jeonghokim.daedongyeojido.domain.announcement.facade.AnnouncementFacade;
+import team.jeonghokim.daedongyeojido.domain.user.domain.User;
+import team.jeonghokim.daedongyeojido.domain.user.facade.UserFacade;
 
 @Service
 @RequiredArgsConstructor
@@ -13,10 +16,17 @@ public class DeleteAnnouncementService {
 
     private final AnnouncementRepository announcementRepository;
     private final AnnouncementFacade announcementFacade;
+    private final UserFacade userFacade;
 
     @Transactional
     public void execute(Long announcementId) {
+        User currentUser = userFacade.getCurrentUser();
         Announcement announcement = announcementFacade.getAnnouncementById(announcementId);
+
+        if (!announcement.getClub().equals(currentUser.getClub())) {
+            throw AnnouncementAccessDeniedException.EXCEPTION;
+        }
+
         announcementRepository.delete(announcement);
     }
 }
